+++ packages/studio-server/src/preview-server/routes/can-update-sequence-props.ts
@@ -10,9 +10,8 @@ import type {
 	TSAsExpression,
 	UnaryExpression,
 } from '@babel/types';
-import type {CanUpdateSequencePropsResponse} from '@remotion/studio-shared';
 import * as recast from 'recast';
-import type {SequenceNodePath} from 'remotion';
+import type {CanUpdateSequencePropsResponse, SequenceNodePath} from 'remotion';
 import type {CanUpdateSequencePropStatus} from 'remotion';
 import {isJsxUnderMapCallback} from '../../codemods/jsx-sequence-context';
 import {parseAst} from '../../codemods/parse-ast';
