+++ packages/studio-server/src/preview-server/sequence-props-watchers.ts
@@ -1,6 +1,5 @@
 import path from 'node:path';
-import type {CanUpdateSequencePropsResponse} from '@remotion/studio-shared';
-import type {SequenceNodePath} from 'remotion';
+import type {CanUpdateSequencePropsResponse, SequenceNodePath} from 'remotion';
 import {installFileWatcher} from '../file-watcher';
 import {waitForLiveEventsListener} from './live-events';
 import {getCachedNodePath, setCachedNodePath} from './node-path-cache';
