+++ packages/studio-shared/src/api-requests.ts
@@ -11,7 +11,7 @@ import type {
 } from '@remotion/renderer';
 import type {HardwareAccelerationOption} from '@remotion/renderer/client';
 import type {_InternalTypes, SequenceSchema} from 'remotion';
-import type {CanUpdateSequencePropStatus} from 'remotion';
+import type {CanUpdateSequencePropsResponse} from 'remotion';
 import type {SequenceNodePath} from 'remotion';
 import type {RecastCodemod, VisualControlChange} from './codemods';
 import type {PackageManager} from './package-manager';
@@ -224,19 +224,6 @@ export type UnsubscribeFromSequencePropsRequest = {
 	nodePath: SequenceNodePath;
 	clientId: string;
 };
-
-export type CanUpdateSequencePropsResponse =
-	| {
-			canUpdate: true;
-			props: Record<string, CanUpdateSequencePropStatus>;
-			nodePath: SequenceNodePath;
-			/** True when the JSX is inside a `.map()` callback (list iteration). */
-			jsxInMapCallback: boolean;
-	  }
-	| {
-			canUpdate: false;
-			reason: string;
-	  };
 
 export type SaveSequencePropsRequest = {
 	fileName: string;
